Split proxy.rs into focused submodule tree
The 3,845-line src/proxy.rs is split into src/proxy/{mod,connect,h3 (renamed
http3),hop_by_hop,http,test_support,websocket}.rs with one transport concern
per file. No behavior change — code moves, visibility tightens, tests follow
their target functions.

- mod.rs: run_proxy, run_proxy_with_limit, Shared, accept loop
- hop_by_hop.rs: format_http_version, parse_connection_tokens,
  is_hop_by_hop_header, HOP_BY_HOP_HEADERS
- connect.rs: handle_connect, tunnel, AlwaysResolves
- http3.rs (named http3, not h3, to avoid shadowing the h3 crate):
  init_h3_endpoint, run_h3_accept_loop, handle_h3_*, emit_h3_protocol_event
- websocket.rs: handle_websocket_upgrade, relay_websocket, message_to_info,
  is_websocket_upgrade, connect_upstream_for_upgrade, ProtocolLintCtx
- http.rs: handle_request, handle_inner_request, handle_http_logic, plus
  build_and_write_transaction lifted from inside handle_http_logic to a
  module-private fn
- test_support.rs (#![cfg(test)]): make_shared_with_cfg,
  make_request_with_headers, boxed_empty, read_capture

Cross-handler utilities are pub(super); inter-module entry points are
pub(super); only run_proxy / run_proxy_with_limit remain pub. The file-level
on the two functions that actually need it.

Public API at lint_http::proxy::* is unchanged.
